refactor: reorganize component structure from pages to routes
Move all page-specific components from `/components/pages/` to `/components/routes/`
to better align with the application's route-based architecture:

- dashboard/* components: forms, modals, navigation, error handling
- fitness.action/* components: exercise displays, set displays, timers, supersets
- media-item/* components: displays, modals, utilities for media item pages

This restructuring improves code organization by grouping components according to
their corresponding route structure, making the codebase more intuitive to navigate.
