+++ apps/frontend/app/components/routes/dashboard/error-boundary.tsx
