+++ apps/frontend/app/components/routes/dashboard/forms/metadata-progress-update/utils/common-elements.tsx
