+++ apps/frontend/app/components/routes/dashboard/navigation/footer.tsx
