+++ apps/frontend/app/components/routes/dashboard/navigation/navigation-config.ts
