+++ apps/frontend/app/components/routes/fitness.action/exercise-display/details-modal.tsx
