+++ apps/frontend/app/components/routes/fitness.action/exercise-display/display.tsx
