+++ apps/frontend/app/components/routes/fitness.action/set-display/display.tsx
