+++ apps/frontend/app/components/routes/fitness.action/set-display/functions.ts
