+++ apps/frontend/app/components/routes/fitness.action/stat-display-and-input.tsx
