+++ apps/frontend/app/routes/_dashboard.media.item.$id._index.tsx
@@ -81,14 +81,14 @@ import {
 import {
 	JUST_WATCH_URL,
 	METADATA_LOTS_WITH_GRANULAR_UPDATES,
-} from "~/components/pages/media-item/constants";
-import { HistoryItem } from "~/components/pages/media-item/displays/history-item";
-import { MetadataCreator } from "~/components/pages/media-item/displays/metadata-creator";
-import { DisplayPodcastEpisode } from "~/components/pages/media-item/displays/podcast-episode";
-import { DisplayShowSeason } from "~/components/pages/media-item/displays/show-season";
-import { VideoIframe } from "~/components/pages/media-item/displays/video-iframe";
-import { MergeMetadataModal } from "~/components/pages/media-item/modals/merge-metadata-modal";
-import { DisplayShowSeasonEpisodesModal } from "~/components/pages/media-item/modals/show-season-episodes-modal";
+} from "~/components/routes/media-item/constants";
+import { HistoryItem } from "~/components/routes/media-item/displays/history-item";
+import { MetadataCreator } from "~/components/routes/media-item/displays/metadata-creator";
+import { DisplayPodcastEpisode } from "~/components/routes/media-item/displays/podcast-episode";
+import { DisplayShowSeason } from "~/components/routes/media-item/displays/show-season";
+import { VideoIframe } from "~/components/routes/media-item/displays/video-iframe";
+import { MergeMetadataModal } from "~/components/routes/media-item/modals/merge-metadata-modal";
+import { DisplayShowSeasonEpisodesModal } from "~/components/routes/media-item/modals/show-season-episodes-modal";
 import { MEDIA_DETAILS_HEIGHT, reviewYellow } from "~/lib/shared/constants";
 import { convertTimestampToUtcString, dayjsLib } from "~/lib/shared/date-utils";
 import {
